Remove obsolete cast to uint16.

+++ js/jsscan.cpp
@@ -993,7 +993,7 @@ NewToken(JSTokenStream *ts, ptrdiff_t adjust)
     tp->pos.begin.index = ts->linepos +
                           PTRDIFF(tp->ptr, ts->linebuf.base, jschar) -
                           ts->ungetpos;
-    tp->pos.begin.lineno = tp->pos.end.lineno = (uint32_t)ts->lineno;
+    tp->pos.begin.lineno = tp->pos.end.lineno = (uint32)ts->lineno;
     return tp;
 }
 
@@ -1085,7 +1085,7 @@ js_GetToken(JSContext *cx, JSTokenStream *ts)
             if (!atom)
                 goto error;
         }
-        tp->pos.end.lineno = (uint16)ts->lineno;
+        tp->pos.end.lineno = (uint32)ts->lineno;
         tp->t_op = JSOP_STRING;
         tp->t_atom = atom;
         goto out;
@@ -1185,7 +1185,7 @@ js_GetToken(JSContext *cx, JSTokenStream *ts)
             atom = TOKENBUF_TO_ATOM();
             if (!atom)
                 goto error;
-            tp->pos.end.lineno = (uint16)ts->lineno;
+            tp->pos.end.lineno = (uint32)ts->lineno;
             tp->t_op = JSOP_STRING;
             tp->t_atom = atom;
             tt = TOK_XMLATTR;
@@ -1449,7 +1449,7 @@ retry:
         atom = TOKENBUF_TO_ATOM();
         if (!atom)
             goto error;
-        tp->pos.end.lineno = (uint16)ts->lineno;
+        tp->pos.end.lineno = (uint32)ts->lineno;
         tp->t_op = JSOP_STRING;
         tp->t_atom = atom;
         tt = TOK_STRING;
@@ -1682,7 +1682,7 @@ retry:
                 if (!atom)
                     goto error;
                 tp->t_atom = atom;
-                tp->pos.end.lineno = (uint16)ts->lineno;
+                tp->pos.end.lineno = (uint32)ts->lineno;
                 goto out;
             }
